pyproject.toml
[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["Nikita Renjov <n.renjov@cft.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyjwt = "^2.8.0"
python-decouple = "^3.8"
pytest = "^8.2.2"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.8"
httpx = "^0.27.0"
kafka-python = "^2.0.2"

[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^0.19.2"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
